+++ internal/templates/admin/locations_templ.go
@@ -772,7 +772,7 @@ func EditLocation(location models.Location, settings models.InstanceSettings, co
 		if templ_7745c5c3_Err != nil {
 			return templ_7745c5c3_Err
 		}
-		for _, block := range blocks.GetRegisteredBlocks() {
+		for _, block := range blocks.GetBlocksForContext(blocks.ContextLocation) {
 			if !block.RequiresValidation() {
 				templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 80, "<div class=\"indicator w-full\"><div class=\"btn btn-outline border-base-content/30 h-auto p-3 tooltip flex flex-col gap-1 items-center rounded-md w-full\" data-tip=\"")
 				if templ_7745c5c3_Err != nil {